pyproject.toml
[tool.poetry]
name = "auth-backend"
version = "0.1.0"
description = ""
authors = ["ptu <ptu@hexdump.it>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.74.1"
python-jose = "^3.3.0"
uvicorn = "^0.17.5"
starlette-exporter = "^0.12.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
